range
range function returns array type after it gets number parameter.  range vs L.range. range is just array. but L.range is iterator type. Exactly latter is not calculated when it is defined. but the former is calculated when defined. in reduce function, array type is translated to a iterator type. So it spends a time more.